Move LogoutViewModel to its own file in production
